Move trait to crate root
